Use asynchronous requests to get subjurisdiction URLs
The URLs scraped from a state's select-county.html page are not the URLs for the county's summary page.  Instead, the pages linked from the select county page use JavaScript or a meta refresh tag to redirect to the county page.  However, if we fetch them, we can figure out the final URL with the identifier that we can scrape from the redirect page.  Still, this means that we have to fetch and scrape one page for each county in a state, which can be slow.

Use the [requests-futures](https://github.com/ross/requests-futures) package to perform multiple asynchronous requests without blocking.  Right now we're using 10 worker threads, but I'm open to something else if this seems to aggressive.

Without asynchronous requests, hitting the real server, ``TestJurisdiction.test_get_subjurisdictions_state()`` takes about 41 seconds to run.  With the async requests, it takes about 6 seconds.

Remove 'clarity_' from a few Jurisdiction method names because it's redundant.  Probably an artifact from the code that used to be in Arkansas' datasource in openelex-core.

Fixes https://github.com/openelections/clarify/issues/2