Call parent class' __new__() instead of __init__()
Calling the parent constructor in our namedtuple subclasses was
breaking in Python 3.4 (and maybe other Python 3 versions).

The tuple part of our objects are immutable, so we need to call
the parent class' __new__() method, not __init__().

Also add 'Base' to the class name argument of the nametuple factory.
It doesn't seem to matter one way or another, but using 'Base'
seems more correct.