Mock summary URL in jurisdiction test
The initialization of the Jurisdiction class makes multiple
external requests in order to confirm that constructed URLs are
valid. If not all of these URLs are mocked or passed through,
responses will raise a ConnectionError and the test will fail.

This hardcodes the URL missing from the mock, but the Jurisdiction
class should be refactored to get at these URLs without hitting
an external server.